Split display implementations into separate files
