+++ platformio.ini
@@ -37,7 +37,6 @@ lib_deps =
 	jsware/AbleButtons@^0.4.0
 	adafruit/Adafruit GFX Library@^1.11.9
 	adafruit/Adafruit SSD1306@^2.5.10
-	pfeerick/elapsedMillis@^1.0.6
 
 [env:esp32dev]
 platform = espressif32
@@ -64,6 +63,5 @@ debug_test = *
 build_flags = -Wp,-w -D PIO_UNIT_TESTING=1
 debug_build_flags = -O0 -g -ggdb	
 lib_deps = 
-	pfeerick/elapsedMillis@^1.0.6
 	google/googletest@^1.15.2
